move v4 client to a dedicated folder
