+++ packages/js/src/fabric/Conversation.ts
@@ -1,7 +1,7 @@
 import { type ConversationEventParams } from '@signalwire/core'
 import { HTTPClient } from './HTTPClient'
 import { WSClient } from './WSClient'
-import { WSClientV4 } from './WSClientV4'
+import { WSClientV4 } from './v4/WSClientV4'
 import type {
   GetConversationsResponse,
   GetMessagesParams,
@@ -224,10 +224,14 @@ export class Conversation {
         missingReturns = chatOnlyMessages.slice(remaining)
       }
 
-      chatMessages = await Promise.all(chatMessages.map(async message => ({
-        ...message,
-        user_name: (await this.httpClient.getAddress({id: message.from_address_id})).display_name
-      })))
+      chatMessages = await Promise.all(
+        chatMessages.map(async (message) => ({
+          ...message,
+          user_name: (
+            await this.httpClient.getAddress({ id: message.from_address_id })
+          ).display_name,
+        }))
+      )
 
       return {
         data: chatMessages as ConversationChatMessage[],
