+++ packages/js/src/fabric/index.ts
@@ -1,17 +1,16 @@
 export * from './SignalWire'
-export * from './SignalWireV4'
 export * from './interfaces/address'
 export * from './interfaces/capabilities'
 export * from './interfaces/conversation'
 export * from './interfaces/device'
 export * from './interfaces/incomingCallManager'
+export * from './v4/SignalWireV4'
 
 export {
   OnlineParams,
   HandlePushNotificationParams,
   HandlePushNotificationResult,
   DialParams,
-  ReattachParams,
 } from './interfaces/wsClient'
 export {
   SignalWireClient,
