+++ packages/js/src/fabric/v4/createWSClientV4.ts
@@ -1,6 +1,6 @@
 import { configureStore } from '@signalwire/core'
 import { SATSessionV4 } from './SATSessionV4'
-import { WSClientOptions } from './interfaces'
+import { WSClientOptions } from '../interfaces'
 
 export const createWSClientV4 = (options: WSClientOptions) => {
   const store = configureStore({
